MIDI file investigation and conversion (Almost there...)
I've spent few weeks trying to understand the structure of MIDI files and to extract only necessary parts: notes.

It turned out a bit tricky since I had to adjust the time tick values according to the changes in tempo. I can of course simply ignore tempo changes, but if I do so, the pace of the music becomes quite weird.

Conversion to text is almost done. I'm planning to write the code for conversion in py script and not in ipynb notebook file.